update controller path to be api/v1/..
- except application controller
- update rotues file added namespace to match url
- Add order_by_name scope to the itemcontroller
