+++ app/controllers/api/v1/users_controller.rb
@@ -1,4 +1,4 @@
-class UsersController < ApplicationController
+class Api::V1::UsersController < ApplicationController
   before_action :authorize_user, only: [:show]
 
   def show
