+++ app/models/item.rb
@@ -3,4 +3,6 @@ class Item < ApplicationRecord
   has_many :users, through: :appointments
 
   validates_presence_of :name
+
+  scope :order_by_name, -> { order(name: :asc) }
 end
